+++ modules/cards/create_new_card/newCardCreate_positive.spec.js
@@ -1,9 +1,9 @@
 import {expect} from 'chai';
 import FlashCardsPage from "../../_PageObjects/FlashCardsPage";
 import LoginPage from "../../_PageObjects/LoginPage";
-import {positive, pageTitle} from "./newFlashCard_data";
+import {positive, pageTitle} from "./newCard_data";
 
-describe('NEW FLASH CARD CREATE', ()=> {
+describe('NEW FLASH CARD CREATE POSITIVE', ()=> {
    before('should login as student', () => {
 // ====== currently logging in as an admin. Waiting for login(role) implementation ==========
         LoginPage.login();
