+++ modules/cards/create_new_card/newCard_data.js
